Ignore v* branches
By default, antora builds master and any v* branch. This caused an issue
with the `VIA` branch which matched the glob and then threw an error
because the nav files were for the same version. This commit makes
antora only look at master for now which should avoid the build issue.